// client/state/Products/products.action-types.ts
export enum ActionTypes {
  FETCH_PRODUCTS_START = 'FETCH_PRODUCTS_START',
  FETCH_PRODUCTS_SUCCESS = 'FETCH_PRODUCTS_SUCCESS',
  FETCH_PRODUCTS_ERROR = 'FETCH_PRODUCTS_ERROR',

  FETCH_TOP_PRODUCTS_START = 'FETCH_TOP_PRODUCTS_START',
  FETCH_TOP_PRODUCTS_SUCCESS = 'FETCH_TOP_PRODUCTS_SUCCESS',
  FETCH_TOP_PRODUCTS_ERROR = 'FETCH_TOP_PRODUCTS_ERROR',

  FETCH_PRODUCT_START = 'FETCH_PRODUCT_START',
  FETCH_PRODUCT_SUCCESS = 'FETCH_PRODUCT_SUCCESS',
  FETCH_PRODUCT_ERROR = 'FETCH_PRODUCT_ERROR',
  FETCH_PRODUCT_RESET = 'FETCH_PRODUCT_RESET',

  DELETE_PRODUCT_START = 'DELETE_PRODUCT_START',
  DELETE_PRODUCT_SUCCESS = 'DELETE_PRODUCT_SUCCESS',
  DELETE_PRODUCT_ERROR = 'DELETE_PRODUCT_ERROR',

  CREATE_PRODUCT_START = 'CREATE_PRODUCT_START',
  CREATE_PRODUCT_SUCCESS = 'CREATE_PRODUCT_SUCCESS',
  CREATE_PRODUCT_ERROR = 'CREATE_PRODUCT_ERROR',

  UPDATE_PRODUCT_START = 'UPDATE_PRODUCT_START',
  UPDATE_PRODUCT_SUCCESS = 'UPDATE_PRODUCT_SUCCESS',
  UPDATE_PRODUCT_ERROR = 'UPDATE_PRODUCT_ERROR',
  UPDATE_PRODUCT_RESET = 'UPDATE_PRODUCT_RESET',

  CREATE_PRODUCT_REVIEW_START = 'CREATE_PRODUCT_REVIEW_START',
  CREATE_PRODUCT_REVIEW_SUCCESS = 'CREATE_PRODUCT_REVIEW_SUCCESS',
  CREATE_PRODUCT_REVIEW_ERROR = 'CREATE_PRODUCT_REVIEW_ERROR',
  CREATE_PRODUCT_REVIEW_RESET = 'CREATE_PRODUCT_REVIEW_RESET',

  FETCH_CATEGORIES_START = 'FETCH_CATEGORIES_START',
  FETCH_CATEGORIES_SUCCESS = 'FETCH_CATEGORIES_SUCCESS',
  FETCH_CATEGORIES_ERROR = 'FETCH_CATEGORIES_ERROR',


}